Adding a new Point class example using glib.
This is a cleaner and a documented example to implement OOP using glib.
Source of this example: http://sgros.blogspot.com/2016/01/few-tips-about-gobject-for-oo.html.

Then I adjuste the main.c and meson.build to include the new files of the point class.

Finally creating a folder Glib Resources to save any useful doc or link for glib.